# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "战鸽数据",
    "description": "Zhange Stats · Steam 游玩统计与圈子成员管理",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/members": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "List Members",
        "description": "只读列表。成员补齐依赖启动 sync_users_and_members / 登录懒加载 ensure_user_member。",
        "operationId": "list_members_api_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOut"
                  },
                  "type": "array",
                  "title": "Response List Members Api Members Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/members/{member_id}": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Get Member",
        "operationId": "get_member_api_members__member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/jobs/members": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "List Job Filter Members",
        "description": "供任务调度页「按用户」下拉：仅含任一签到平台已绑定的成员。",
        "operationId": "list_job_filter_members_api_settings_jobs_members_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobMemberOptionOut"
                  },
                  "type": "array",
                  "title": "Response List Job Filter Members Api Settings Jobs Members Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/steam/members/{member_id}": {
      "get": {
        "tags": [
          "steam"
        ],
        "summary": "Steam Member Stats",
        "operationId": "steam_member_stats_api_steam_members__member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPlayStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kujiequ/status": {
      "get": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Status",
        "operationId": "kujiequ_status_api_kujiequ_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "include_roles",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Roles"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kujiequ/logs": {
      "get": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Logs",
        "description": "兼容占位：历史列表已弃用；今日状态见 status（读 *_checkin_logs 缓存）。",
        "operationId": "kujiequ_logs_api_kujiequ_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckinLogOut"
                  },
                  "title": "Response Kujiequ Logs Api Kujiequ Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kujiequ/bind/token": {
      "post": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Bind Token",
        "operationId": "kujiequ_bind_token_api_kujiequ_bind_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KujiequBindTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/kujiequ/bind/sms/send": {
      "post": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Bind Sms Send",
        "operationId": "kujiequ_bind_sms_send_api_kujiequ_bind_sms_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KujiequBindSmsSendRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequBindSmsSendResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/kujiequ/bind/sms": {
      "post": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Bind Sms",
        "operationId": "kujiequ_bind_sms_api_kujiequ_bind_sms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KujiequBindSmsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/kujiequ/bind": {
      "delete": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Unbind",
        "operationId": "kujiequ_unbind_api_kujiequ_bind_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequStatusOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Patch Bind",
        "operationId": "kujiequ_patch_bind_api_kujiequ_bind_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KujiequBindUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KujiequStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/kujiequ/checkin": {
      "post": {
        "tags": [
          "kujiequ"
        ],
        "summary": "Kujiequ Checkin",
        "operationId": "kujiequ_checkin_api_kujiequ_checkin_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/napcat/groups/{group_id}/members": {
      "get": {
        "tags": [
          "napcat"
        ],
        "summary": "List Group Members",
        "operationId": "list_group_members_api_napcat_groups__group_id__members_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "强制不走 NapCat 缓存",
              "default": false,
              "title": "Force"
            },
            "description": "强制不走 NapCat 缓存"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NapCatGroupMembersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "game_code": {
            "type": "string",
            "title": "Game Code"
          },
          "game_name": {
            "type": "string",
            "title": "Game Name"
          },
          "role_uid": {
            "type": "string",
            "title": "Role Uid"
          },
          "role_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "channel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label",
            "default": ""
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "awards_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards Text"
          },
          "checkin_date": {
            "type": "string",
            "title": "Checkin Date"
          },
          "checked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Checked At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "game_code",
          "game_name",
          "role_uid",
          "status",
          "checkin_date",
          "checked_at"
        ],
        "title": "CheckinLogOut",
        "description": "签到记录公共结构（森空岛 / 塔吉多等共用）。"
      },
...
      "CheckinResponse": {
        "properties": {
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ok"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CheckinResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobMemberOptionOut": {
        "properties": {
          "member_id": {
            "type": "integer",
            "title": "Member Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "label"
        ],
        "title": "JobMemberOptionOut"
      },
...
      "KujiequBindSmsRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "captcha": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Captcha"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "captcha"
        ],
        "title": "KujiequBindSmsRequest"
      },
      "KujiequBindSmsSendRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 6,
            "title": "Phone"
          },
          "gee_test_data": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192
              },
              {
                "type": "null"
              }
            ],
            "title": "Gee Test Data"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "KujiequBindSmsSendRequest"
      },
      "KujiequBindSmsSendResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "验证码已发送"
          },
          "need_geetest": {
            "type": "boolean",
            "title": "Need Geetest",
            "default": false
          },
          "captcha_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captcha Id"
          }
        },
        "type": "object",
        "title": "KujiequBindSmsSendResponse"
      },
      "KujiequBindTokenRequest": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 8,
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "KujiequBindTokenRequest"
      },
      "KujiequBindUpdate": {
        "properties": {
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 23.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 59.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          }
        },
        "type": "object",
        "title": "KujiequBindUpdate"
      },
...
      "KujiequStatusOut": {
        "properties": {
          "bound": {
            "type": "boolean",
            "title": "Bound"
          },
          "auto_checkin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Checkin"
          },
          "checkin_hour": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Hour"
          },
          "checkin_minute": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkin Minute"
          },
          "phone_mask": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Mask"
          },
          "bound_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bound At"
          },
          "last_checkin_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin At"
          },
          "last_checkin_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Date"
          },
          "last_checkin_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Ok"
          },
          "last_checkin_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Checkin Summary"
          },
          "token_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Ok"
          },
          "token_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Error"
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/KujiequRoleOut"
            },
            "type": "array",
            "title": "Roles"
          },
          "today_results": {
            "items": {
              "$ref": "#/components/schemas/CheckinResultItem"
            },
            "type": "array",
            "title": "Today Results"
          },
          "today_logs": {
            "items": {
              "$ref": "#/components/schemas/CheckinLogOut"
            },
            "type": "array",
            "title": "Today Logs"
          }
        },
        "type": "object",
        "required": [
          "bound"
        ],
        "title": "KujiequStatusOut"
      },
...
      "MemberOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nickname": {
            "type": "string",
            "title": "Nickname"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nickname",
          "avatar_url",
          "user_id",
          "joined_at"
        ],
        "title": "MemberOut"
      },
...
      "MemberPlayStatsResponse": {
        "properties": {
          "member": {
            "$ref": "#/components/schemas/MemberPlayMember"
          },
          "week_play_seconds": {
            "type": "integer",
            "title": "Week Play Seconds"
          },
          "month_play_seconds": {
            "type": "integer",
            "title": "Month Play Seconds"
          },
          "session_count": {
            "type": "integer",
            "title": "Session Count"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/PlayTrendPoint"
            },
            "type": "array",
            "title": "Trend"
          },
          "recent_sessions": {
            "items": {
              "$ref": "#/components/schemas/SteamSessionBrief"
            },
            "type": "array",
            "title": "Recent Sessions"
          }
        },
        "type": "object",
        "required": [
          "member",
          "week_play_seconds",
          "month_play_seconds",
          "session_count",
          "trend",
          "recent_sessions"
        ],
        "title": "MemberPlayStatsResponse"
      },
...
      "NapCatGroupMembersResponse": {
        "properties": {
          "group_id": {
            "type": "string",
            "title": "Group Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/NapCatGroupMemberOut"
            },
            "type": "array",
            "title": "Members"
          },
          "site_bound_count": {
            "type": "integer",
            "title": "Site Bound Count"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "members",
          "site_bound_count"
        ],
        "title": "NapCatGroupMembersResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}